Rename producer's utils library to be compliant with c0018.001 dev-cycle
