+++ producer/telemetries/test_client.py
@@ -5,8 +5,8 @@ from lsst.ts import salobj
 
 # from mock import patch, mock_open
 
-from .. import test_utils
-from .. import utils
+import test_utils
+import producer_utils
 
 STD_TIMEOUT = 15  # timeout for command ack
 SHOW_LOG_MESSAGES = False
@@ -63,7 +63,7 @@ class TestTelemetriesClient(test_utils.WSClientTestCase):
             expected_stream = {
                 p: {
                     "value": getattr(tel_scalars, p),
-                    "dataType": utils.get_data_type(getattr(tel_scalars, p)),
+                    "dataType": producer_utils.get_data_type(getattr(tel_scalars, p)),
                     "units": f"{self.remote.tel_scalars.metadata.field_info[p].units}",
                 }
                 for p in tel_parameters
@@ -73,11 +73,11 @@ class TestTelemetriesClient(test_utils.WSClientTestCase):
             while True:
                 response = await websocket.recv()
                 message = json.loads(response)
-                stream_exists = utils.check_event_stream(
+                stream_exists = producer_utils.check_event_stream(
                     message, "telemetry", "Test", self.index, "scalars"
                 )
                 if stream_exists:
-                    stream = utils.get_event_stream(
+                    stream = producer_utils.get_event_stream(
                         message, "telemetry", "Test", self.index, "scalars"
                     )
                     break
@@ -150,7 +150,7 @@ class TestTelemetriesClient(test_utils.WSClientTestCase):
             expected_stream = {
                 p: {
                     "value": getattr(tel_scalars, p),
-                    "dataType": utils.get_data_type(getattr(tel_scalars, p)),
+                    "dataType": producer_utils.get_data_type(getattr(tel_scalars, p)),
                     "units": f"{self.remote.tel_scalars.metadata.field_info[p].units}",
                 }
                 for p in tel_parameters
@@ -160,11 +160,11 @@ class TestTelemetriesClient(test_utils.WSClientTestCase):
             while True:
                 response = await websocket.recv()
                 message = json.loads(response)
-                stream_exists = utils.check_event_stream(
+                stream_exists = producer_utils.check_event_stream(
                     message, "telemetry", "Test", self.index, "scalars"
                 )
                 if stream_exists:
-                    stream = utils.get_event_stream(
+                    stream = producer_utils.get_event_stream(
                         message, "telemetry", "Test", self.index, "scalars"
                     )
                     break
